--- a/maps/biter_battles_v2/terrain.lua
+++ b/maps/biter_battles_v2/terrain.lua
@@ -325,16 +325,20 @@ local function draw_biter_area(surface, left_top_x, left_top_y)
 	for x = 0, 31, 1 do
 		for y = 0, 31, 1 do
 			local position = { x = left_top_x + x, y = left_top_y + y }
+			if not Functions.is_biter_area(position, true) then
+				goto d_b_a
+			end
+
 			-- + 1, because lua has 1-based indices
 			local grid_p_x = ((position.x + seed) % biter_texture.width) + 1
 			local grid_p_y = ((position.y + seed) % biter_texture.height) + 1
 			local id = biter_texture.grid[grid_p_x][grid_p_y]
 			local name = biter_texture.map[id]
-			if Functions.is_biter_area(position,true) then
-				out_of_map[i] = {name = "out-of-map", position = position}
-				tiles[i] = {name = name, position = position}
-				i = i + 1
-			end
+			out_of_map[i] = {name = "out-of-map", position = position}
+			tiles[i] = {name = name, position = position}
+			i = i + 1
+
+			::d_b_a::
 		end
 	end