… maps/biter_battles_v2/biter_texture.lua