### Brief description of the changes:
This patchset optimizes biter terrain generation by removing resource/math intensive simplex noise computations. At present, we run 1024 noise functions per each chunk in biter area. The side effect of this change is that pattern of biter area had to be changed to other 2D seamless texture. There is no problem changing the texture as it comes at no performance impact.

Speed at which 1024 tiles were generated previously (tiles only, not structures etc):
max: 27.3ms
min: 6.9ms
mean: 10.5ms

When this solution is applied:
max: 8.8ms
min: 1.0ms
mean: 2.6ms

I took the measurements after map was fully loaded and just idling. Then I navigated towards nests and kept pushing for few seconds. I then took sample from last 10s and computed metrics based on that.

Because we're constrained to 16ms+- per each tick, new solution frees up lots of time to handle other events like threat farming, which typically triggers map generation in biter area.

![1](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/59515242/c9bbbcbd-372c-4128-b305-7e49d1f81540)
(Screenshot modifies colors, it looks less saturated in game)

### How texture is generated
First I found seamless cartoonish 2D texture, I processed it to highlight the edges, turned into grayscale, manipulated contrast and resized image to 128x128. As values on grayscale are between 0 - 127 I was able to map different shade ranges to different shades of dirt tiles.
The mapping was performed by separate program I wrote for this occasion (not sure where to upload it). Output of the mapping is complete lua script loaded as part of scenario that contains hardcoded values per each position within 128x128 2D lua array.
![texture](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/59515242/40c4e940-17a4-4612-873e-6f31ef160ab4)

### Tested Changes:
- [X] I've tested the changes locally
- [ ] I've tested the changes online

Here is the vote about this update, as the visual was edited, just in case... : 
https://discord.com/channels/823696400797138974/823771211421974579/1162712684345688135