@Ragnarok77-factorio 2023-10-15
For now I just tested it, didn't find issue or error. 
I didn't and won't try though to be honest to deeply test and compare perf on my side too, trusting his results, but if anyone else is motivated to do the check , feel free, here are the screenshots : 
Before : 
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/95619848/f03dbebd-4b62-4dd1-9b49-65cfc5961bfc)
After : 
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/95619848/60e696fa-5000-477f-8c2f-4b9d13c29d26)
Before : 
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/95619848/670a0004-4cfe-4361-a892-8e3be59e18ab)
After : 
![image](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/95619848/b68a2183-a6e9-4397-bc6f-4011578473a6)