@Ragnarok77-factorio 2023-10-18
I tested it locally too, no issue