@cogito123 2023-10-18
I applied new 512x512 grid (previous one was 128x128) which is about 536KB in size. It's less repetitive at larger scale and looks better at closer look.
![texture2](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/59515242/8f65a13b-acee-4256-b575-bee4e0a4d934)
![minimap](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/59515242/a968eee7-3465-456e-8d0a-d0ff28f32ec0)
![img](https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/assets/59515242/7581939c-adc8-48ee-9600-9abcd650f8d2)