@Ragnarok77-factorio 2023-10-19
It's a lot worse visually that I thought, tested it, no issue, ok for me.