@Ragnarok77-factorio 2023-10-27
No one complained about the visual of it. fine for me.